requirements.txt
bcrypt==4.1.2
beautifulsoup4==4.9.3
blinker==1.7.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.9  # Or whatever the latest version is
requests==2.26.0
soupsieve==2.2.1
SQLAlchemy==2.0.27
typing_extensions==4.10.0
urllib3==1.26.6
